moved movie model
