+++ libs/user/src/lib/user/+state/index.ts
@@ -1,2 +1,1 @@
-export * from './user.model';
 export * from './user.service';
